Cleanup + resolve
