+++ library/management/chat/src/main/java/org/quiltmc/qsl/chat/api/types/ImmutableS2CProfileIndependentMessage.java
@@ -43,6 +43,7 @@ public final class ImmutableS2CProfileIndependentMessage extends MutableAbstract
 				})
 		);
 	}
+
 	public ImmutableS2CProfileIndependentMessage(PlayerEntity player, boolean isOnClientSide, Text message, MessageType.Parameters messageType) {
 		super(player, isOnClientSide);
 		this.message = message;
