+++ library/management/chat/src/main/java/org/quiltmc/qsl/chat/api/types/ImmutableS2CSystemMessage.java
@@ -45,6 +45,7 @@ public final class ImmutableS2CSystemMessage extends ImmutableAbstractMessage<Im
 		this.content = content;
 		this.overlay = overlay;
 	}
+
 	@Override
 	public @NotNull EnumSet<QuiltMessageType> getTypes() {
 		return InternalQuiltChatApiUtil.s2cType(QuiltMessageType.SYSTEM, isOnClientSide);
